requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
black==25.1.0
cffi==1.17.1
cfgv==3.4.0
click==8.1.8
cryptography==44.0.2
decorator==5.2.1
distlib==0.3.9
Django==5.1.7
django-environ==0.12.0
django-extensions==3.2.3
djangorestframework==3.15.2
drf-spectacular==0.28.0
executing==2.2.0
filelock==3.17.0
gunicorn==23.0.0
identify==2.6.8
inflection==0.5.1
ipython==9.0.1
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.6
pre_commit==4.1.0
prompt_toolkit==3.0.50
psycopg==3.2.5
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
rpds-py==0.23.1
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
uritemplate==4.1.1
virtualenv==20.29.3
wcwidth==0.2.13
psycopg2-binary==2.9.10
requests~=2.32.3
beautifulsoup4~=4.13.3
black==25.1.0